requirements.txt
discord.py>=2.1
localStoragePy>=0.2.3
